'edit'

+++ Sheets/Sheet2/Solution/src/main/java/solution/Vector.java
@@ -19,7 +19,7 @@ class Vector
       y = 0;
    }
 
-   public Vector(int x, int y)
+   public Vector(double x, double y)
    {
       this.x = x;
       this.y = y;
